PWM support now works using the SH-2 driver from Knuckles Chaotix.
I had issues converting the code to assemble correctly with GAS, so it builds with asmsh as  a binary blob, gets converted to a .h by bin2h and then is included in the slave C code.
The slave C code then copies the driver to Cache and runs it from there.

Green Hill Zone music has been replaced with the music for Door Into Summer for testing purposes. No other tracks work with PWM yet...